requirements.txt
streamlit>=1.32.0
groq>=0.4.2
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
easyocr>=1.7.1
openai-whisper>=20231117
sympy>=1.12
python-dotenv>=1.0.0
Pillow>=10.0.0
soundfile>=0.12.1
librosa>=0.10.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
